Splited tests from test_qulacs_multicpu.py
