+++ python/tests/multi_cpu/conftest.py
@@ -1,0 +1,16 @@
+import qulacs
+
+multicpu = False
+if qulacs.check_build_for_mpi():
+    try:
+        from mpi4py import MPI
+
+        mpicomm = MPI.COMM_WORLD
+        if mpicomm.Get_rank() == 0:
+            print("Test with MPI. size=", mpicomm.Get_size())
+
+        multicpu = True
+    except Exception as e:
+        print(e)
+        print("To use multi-cpu, the mpi4py library is required.")
+        exit()
