+++ python/tests/multi_cpu/test_state_multi_cpu.py
@@ -1,0 +1,54 @@
+import unittest
+
+import numpy as np
+
+import qulacs
+
+
+class TestQuantumState(unittest.TestCase):
+    def setUp(self):
+        self.n = 6
+        self.state = qulacs.QuantumState(self.n)
+        self.state_multi = qulacs.QuantumState(self.n, True)
+        self.dim_all = 2**self.n
+        self.dim = 2**self.n
+        if multicpu:
+            self.mpirank = mpicomm.Get_rank()
+            self.mpisize = mpicomm.Get_size()
+        else:
+            self.mpirank = 0
+            self.mpisize = 1
+        if self.state_multi.get_device_name() == "multi-cpu":
+            self.dim //= self.mpisize
+
+    def tearDown(self):
+        del self.state
+        del self.state_multi
+
+    def test_state_dim(self):
+        vector = self.state_multi.get_vector()
+        self.assertEqual(len(vector), self.dim, msg="check vector size")
+
+    def test_zero_state(self):
+        self.state_multi.set_zero_state()
+        vector = self.state_multi.get_vector()
+        vector_ans = np.zeros(self.dim)
+        if self.state_multi.get_device_name() == "cpu" or self.mpirank == 0:
+            vector_ans[0] = 1.0
+        self.assertTrue(
+            ((vector - vector_ans) < 1e-10).all(), msg="check set_zero_state"
+        )
+
+    def test_comp_basis(self):
+        pos = 0b010100
+        self.state_multi.set_computational_basis(pos)
+        vector = self.state_multi.get_vector()
+        vector_ans = np.zeros(self.dim)
+        if (
+            self.state_multi.get_device_name() == "cpu"
+            or self.mpirank == pos // self.dim
+        ):
+            vector_ans[pos % self.dim] = 1.0
+        self.assertTrue(
+            ((vector - vector_ans) < 1e-10).all(), msg="check set_computational_basis"
+        )
